+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - hnefatafl-client: Hide the password.
-- hnefatafl-client: do logging.
+- hnefatafl-client: Do logging.
 
 ## [0.4.1] - 2025-02-04