+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- hnefatafl-client: Improve the GUI.
 - Resign if you can't generate a move.
 
 ## [0.4.1] - 2025-02-04