+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2025-02-04
+
 ### Added
 
 - Make a users screen.
@@ -31,5 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - hnefatafl-client: Make users and games scrollable.
 - Make all the features dependencies.
 
-[unreleased]: https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v0.4.1...main
+[unreleased]: https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v0.5.0...main
+[0.5.0]: https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v0.4.0...v0.4.1